+++ src/js/customer/customers.component.ts
@@ -39,6 +39,7 @@ export class CustomersComponent implements OnInit {
     }
     
     remove(device:Customer, evt:any) {
+        if(!confirm('确定要删除吗')) return;
         this.service.delete(device.id).then(res => {
             var ret = res.json();
             if(ret.code === 'success') {
@@ -84,7 +85,8 @@ export class CustomersComponent implements OnInit {
         this.service.create(this.SelectedModel).then(res => {
             var ret = res.json();
             if(ret.code==='success') {
-                _this.models.unshift(ret.result as Customer);
+                _this.pageChange(1);
+                //_this.models.unshift(ret.result as Customer);
                 btn.click();
             } else {
                 _this.popBy('#btnCreate', ret.message,'');
